Rename api-paste and policy configuration files
The api-paste file was incorrectly named api_paste.ini when implemented
on the designate server, and therefore wasn't being used.

The policy.json template file was renamed to policy.json.j2 to keep
consistent with conventions in OpenStack-Ansible. This eases the process
for updating the files automatically through convenience scripts.