Simplify test inventory
The previous inventory was 'dynamic' in that the host/group
information was added through a playbook. This then required
the playbook implementing the inventory to be executed if
the deployer wished to test any single playbook.

This implements a static inventory and group_vars to achieve
the same result, simplifying the test implementation and
providing greater flexibility when executing the test playbooks.

Change-Id: Ia5b35efd42955cc840990e798eee96a79926b134